[Silver I] Title: 카잉 달력, Time: 24 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/6064. 카잉 달력/카잉 달력.cc
@@ -39,7 +39,7 @@ int gcd(int a, int b) {
 	return a;
 }
 
-long long lcm(long long a, long long b) {
+int lcm(int a, int b) {
 	return a * b / gcd(a, b);
 }
